refactor(frontend): group Payload source under src/cms/
Payload's config, collections, globals, migrations, and seed script
were scattered directly under src/, mixed in with app code. Moved
them all under src/cms/ so Payload-specific source is grouped in one
place, distinct from the Next.js app and shared utilities.

payload-types.ts stays at the top-level src/ since @/payload-types is
consumed by 3 files through that fixed alias path, and moving it would
have meant updating every consumer for no benefit.

Normalized the 4 collections that used relative ../lib/... imports to
the @/lib/... alias, matching the 3 that already did, so the import
style is consistent across all collections regardless of nesting depth.

Updated every path that pointed at the old locations: tsconfig's
@payload-config alias, eslint's ignore glob, package.json's seed
script, knip's entry/project exclusions, and the labeler.yml glob used
for PR auto-labeling (also folding in globals/, which the old
three-path list had been missing).
