+++ .github/labeler.yml
@@ -5,9 +5,7 @@
 "domain: backend":
   - changed-files:
       - any-glob-to-any-file:
-          - "apps/frontend/src/payload.config.ts"
-          - "apps/frontend/src/collections/**"
-          - "apps/frontend/src/migrations/**"
+          - "apps/frontend/src/cms/**"
           - "apps/frontend/src/app/(payload)/**"
 
 "domain: shared":
