+++ apps/frontend/CLAUDE.md
@@ -7,7 +7,8 @@ This agent works ONLY on `apps/frontend/`. Never modify files outside this direc
 ## Architecture
 
 - Next.js 16 (App Router) + React 19 + TypeScript
-- Payload 3 mounted in this same app: admin at `/admin`, API at `/api/*`
+- Payload 3 mounted in this same app: admin at `/admin`, API at `/api/*`, all
+  Payload-specific source grouped under `src/cms/`
 - Tailwind CSS 4 for styling (no CSS modules)
 - Zustand available for UI/client state (no store needed so far)
 - Server state is read directly in server components and cached server-side — see
@@ -28,9 +29,14 @@ src/
 │   ├── (payload)/      # Payload admin + API — GENERATED, do not edit
 │   └── globals.css     # Tokens + import list, one entry point
 ├── styles/             # Stylesheets shared by several routes
-├── collections/        # Payload collections (users, media)
-├── migrations/         # Payload migrations — reviewed and versioned
-├── payload.config.ts   # Payload configuration
+├── cms/                # Everything Payload-specific
+│   ├── payload.config.ts # Payload configuration
+│   ├── collections/      # Payload collections (users, media, ...)
+│   ├── globals/           # Payload globals (site identity, profile)
+│   ├── migrations/       # Payload migrations — reviewed and versioned
+│   ├── seed/              # `pnpm seed` script and its content
+│   └── open-graph*.ts    # Link preview fetch + hook, used by collections only
+├── payload-types.ts    # Generated by Payload — GENERATED, do not edit
 ├── components/         # Components used by two or more routes
 └── lib/                # Utils, data access, constants
 ```
