+++ apps/frontend/eslint.config.cjs
@@ -4,12 +4,12 @@
 const { createNextConfig } = require('@portfolio/eslint-config/next')
 
 module.exports = [
-  // Fichiers générés par la CLI Payload : toute correction manuelle serait
-  // écrasée à la prochaine génération. Les migrations en font partie — leur
-  // contenu est produit par `migrate:create` — mais elles restent relues et
-  // versionnées, car elles décrivent le schéma de la base.
+  // Files generated by the Payload CLI: any manual fix would be overwritten
+  // at the next generation. Migrations are part of this — their content is
+  // produced by `migrate:create` — but they stay reviewed and versioned,
+  // since they describe the database schema.
   {
-    ignores: ['src/app/(payload)/**', 'src/payload-types.ts', 'src/migrations/**'],
+    ignores: ['src/app/(payload)/**', 'src/payload-types.ts', 'src/cms/migrations/**'],
   },
   ...createNextConfig({ tsconfigRootDir: __dirname }),
 ]
