+++ apps/frontend/src/cms/collections/projects.ts
@@ -1,5 +1,6 @@
-import { CONTENT_TAGS, revalidateCollection } from '../lib/content-cache'
-import { withOpenGraphPreview } from '../lib/open-graph-hook'
+import { CONTENT_TAGS, revalidateCollection } from '@/lib/content-cache'
+
+import { withOpenGraphPreview } from '../open-graph-hook'
 
 import type { CollectionConfig } from 'payload'
