+++ apps/frontend/src/cms/payload.config.ts
@@ -30,7 +30,9 @@ export default buildConfig({
   editor: lexicalEditor(),
   secret: process.env.PAYLOAD_SECRET ?? '',
   typescript: {
-    outputFile: path.resolve(dirname, 'payload-types.ts'),
+    // Generated types live at src/payload-types.ts, one level up from this
+    // config, so they resolve through the plain @/payload-types alias.
+    outputFile: path.resolve(dirname, '..', 'payload-types.ts'),
   },
   db: postgresAdapter({
     pool: {
