+++ apps/frontend/tsconfig.json
@@ -20,7 +20,7 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@payload-config": ["./src/payload.config.ts"]
+      "@payload-config": ["./src/cms/payload.config.ts"]
     }
   },
   "include": [
