+++ knip.config.ts
@@ -7,7 +7,7 @@ const config: KnipConfig = {
       // plugins: only the entry points it cannot infer are listed here.
       entry: [
         // Run through `pnpm seed`, not imported by the app.
-        "src/seed/index.ts",
+        "src/cms/seed/index.ts",
       ],
       // Invoked by `.lintstagedrc.js` as a workspace binary rather than imported,
       // so Knip cannot trace it. The dependency is real: dropping it would break
@@ -31,7 +31,7 @@ const config: KnipConfig = {
         // replaces the whole set, and these exclusions are what keep the
         // Payload-generated files out of the report. They are generated by the
         // CLI, not hand-written, so Knip has nothing useful to say about them.
-        project: ["!src/migrations/**", "!src/payload-types.ts"],
+        project: ["!src/cms/migrations/**", "!src/payload-types.ts"],
       },
     },
     "packages/shared": {},
